Update README.md
V0 of this project was nice, but I will be reviewing the material and starting over from scratch to build out a project structured and organized in a better way. Recently, I have stumbled upon some interesting methods in research across different domains that I want to apply to this problem space to see if anything changes. V1 will become a side-project I'll iterate over time and hopefully lead to some cool findings. Initial ideas subject to change are performing more analysis on the sequence of characters in an attempt to extract some form of unique information and not solely rely on neural networks or machine learning frameworks to yield results. I will seek to demonstrate depth and understanding of the malware domain generation algorithm space without leveraging machine learning as much as possible aka not just throwing into a model and hoping for the best.